refactor: добавляет отдельный модуль под сообщения отправки данных на сервер
